// app/types/TvShowType.ts
type TvShowType = {
    adult: boolean
    backdrop_path: string
    first_air_date: string
    origin_country: string[]
    genre_ids: number[]
    id: number
    name: string
    media_type: string
    original_language: string
    original_name: string
    overview: string
    popularity: number
    poster_path: string
    vote_average: number
    vote_count: number
}

export default TvShowType